Add isdeifined function and comment out exist go script testing by using EvalComplex Linker
